refactor: Break out `ConsentsConfirmation` into it's own component
